requirements.txt
requests==2.22.0
Telethon==1.11.3
GitPython==3.1.0
tinydb==3.15.2
